Allow maps to set max players and allow dynamic player arrays
This allows maps to set higher number of players if there is enough colors.
These changes leverage the need to change replay MD format in this version (3.0.8) to also introduce dynamic players.

NETWORK_PLAYERS_MAX is still used for server related stuff but should be converted to dynamic len in future.